added terminal lib to move cursor and print in terminal easily
