+++ lecteur.py
@@ -1,4 +1,4 @@
-#made by sand 
+#made by sand
 class App:
     def __init__( self ):
         self.init_main()
@@ -17,10 +17,10 @@ class App:
     from libs.handmade._command import init_command,sort_command,edit_command         
     from libs.handmade._param import init_param,get_param,write_param,reset 
     from libs.handmade._files import init_file,get_file,select_dir,switch_dir,find_file,check_adress,change_main_path,mani_file
-    from libs.handmade._image import init_image,get_img,display_img,select_img,display,help_menu,out,ask,ask_list,show_list,change_confirmation
+    from libs.handmade._image import init_image,get_img,display_img,select_img,help_menu,out,ask,ask_list,show_list,change_confirmation
     from libs.handmade._download import init_download,yt_search,dl_yt_playlist  
     from libs.handmade._song import init_song,choose_song,load_songs,play_song,play_last,historic,select,play
-    from libs.handmade._main import init_main,main,update,get_input,load_all,wind,check_time
+    from libs.handmade._main import init_main,main,update,get_input,load_all,wind,check_time,display
     from libs.handmade._letter import init_letter,suspend,a_f,b_f,c_f,d_f,e_f,g_f,h_f,i_f,j_f,l_f,m_f,n_f,o_f,p_f,q_f,r_f,s_f,t_f,u_f,v_f,w_f,x_f,y_f,z_f,plus_f,minus_f,dl_f
 app = App()
 app.main()
